Create manifest.yml
This is a first cut of the manifest for tsrc to use to build up a Yocto metadata set.  6 of one, half dozen of another.  While there's less convenience features with URIs like with Android's repo tool, it's simpler and a bit easier to follow and does much more precisely what we wanted with Yocto.  It's not even close to done, but I wanted something in the repo to work with and build up/test with.